+++ backend/python/app/models/__init__.py
@@ -1,15 +1,23 @@
+# postgresql {
 from flask_sqlalchemy import SQLAlchemy
+# } postgresql
+# mongodb {
 from mongoengine import connect
 
+# } mongodb
+
+# postgresql {
 db = SQLAlchemy()
 erase_db_and_sync = True
 
 
+# } postgresql
 def init_app(app):
     from .entity import Entity
     from .user_pg import User
 
     app.app_context().push()
+    # postgresql {
     db.init_app(app)
 
     if erase_db_and_sync:
@@ -19,6 +27,8 @@ def init_app(app):
 
         # recreate tables
         db.create_all()
-
+    # } postgresql
+    # mongodb {
     # connect to MongoDB
     connect(host=app.config["MONGODB_URL"])
+    # } mongodb
