+++ backend/typescript/models/entity.model.ts
