+++ backend/typescript/models/index.ts
@@ -4,7 +4,7 @@ import { Sequelize } from "sequelize-typescript";
 
 export const sequelize = new Sequelize(
   `postgres://${process.env.POSTGRES_USER}:${process.env.POSTGRES_PASSWORD}@${process.env.DB_HOST}:5432/${process.env.POSTGRES_DB}`,
-  { models: [path.join(__dirname, "/*.pgmodel.ts")] },
+  { models: [path.join(__dirname, "/*.model.ts")] },
 );
 
 // } postgresql
