+++ frontend/src/App.tsx
@@ -1,19 +1,29 @@
 import "bootstrap/dist/css/bootstrap.min.css";
+// auth {
 import React, { useState } from "react";
+// } auth
+// no-auth {
+import React from "react";
+// } no-auth
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
+// auth {
 import Login from "./components/auth/Login";
 import PrivateRoute from "./components/auth/PrivateRoute";
 import CreatePage from "./components/pages/CreatePage";
 import Default from "./components/pages/Default";
 import DisplayPage from "./components/pages/DisplayPage";
+// } auth
 import NotFound from "./components/pages/NotFound";
+// auth {
 import UpdatePage from "./components/pages/UpdatePage";
 import AUTHENTICATED_USER_KEY from "./constants/AuthConstants";
 import AuthContext, { AuthenticatedUser } from "./contexts/AuthContext";
 import { getLocalStorageObj } from "./utils/LocalStorageUtils";
+// } auth
 
 const App = () => {
+  // auth {
   const currentUser: AuthenticatedUser = getLocalStorageObj(
     AUTHENTICATED_USER_KEY,
   );
@@ -22,19 +32,26 @@ const App = () => {
     currentUser,
   );
 
+  // } auth
   return (
+    // auth {
     <AuthContext.Provider value={{ authenticatedUser, setAuthenticatedUser }}>
+    // } auth
       <Router>
         <Switch>
+          // auth {
           <PrivateRoute exact path="/" component={Default} />
           <Route exact path="/login" component={Login} />
           <PrivateRoute exact path="/entity/create" component={CreatePage} />
           <PrivateRoute exact path="/entity/update" component={UpdatePage} />
           <PrivateRoute exact path="/entity" component={DisplayPage} />
+          // } auth
           <Route exact path="*" component={NotFound} />
         </Switch>
       </Router>
+    // auth {
     </AuthContext.Provider>
+    // } auth
   );
 };
